requirements.txt
# Core ML
torch>=2.0
torchvision>=0.15
timm>=0.9.0

# Data & utilities
numpy
pillow
pandas

# Training helpers
tqdm

# Metrics & plotting
scikit-learn
matplotlib
seaborn

# Augmentations (for robustness experiments later)
albumentations>=1.3.0
